refactor: extract add node list
