+++ src/config/designer/ReactFlowCanvas.tsx
@@ -1,9 +1,9 @@
 import React, {
-  useState,
   useCallback,
   useMemo,
   useEffect,
   useRef,
+  useState,
 } from "react";
 import {
   ReactFlow,
@@ -26,26 +26,19 @@ import {
 import "@xyflow/react/dist/style.css";
 import {
   Box,
-  Text,
   ActionIcon,
-  Button,
-  Stack,
-  Transition,
-  Paper,
   Tooltip,
-  ScrollArea,
   useComputedColorScheme,
 } from "@mantine/core";
-import { IconPlus, IconLayoutGrid } from "@tabler/icons-react";
-import { WorkflowNode, WorkflowConnection } from "@/types/workflow";
+import { IconLayoutGrid } from "@tabler/icons-react";
+import { WorkflowNode, WorkflowConnection, NodeType } from "@/types/workflow";
 import dagre from "@dagrejs/dagre";
-import { ActionRegistry } from "@/services/actionRegistry";
-import { TriggerRegistry } from "@/services/triggerRegistry";
 import { initializeTriggers } from "@/services/triggerInitializer";
 import { initializeActions } from "@/services/actionInitializer";
 import CanvasNode from "./CanvasNode";
 import { useHotkeys } from "@mantine/hooks";
 import { NotificationService } from "@/services/notification";
+import AddNodeList from "./AddNodeList";
 
 interface ReactFlowCanvasProps {
   nodes: WorkflowNode[];
@@ -80,9 +73,9 @@ const ReactFlowCanvasInner: React.FC<ReactFlowCanvasProps> = ({
   errors,
 }) => {
   const colorScheme = useComputedColorScheme();
-  const [showNodePalette, setShowNodePalette] = useState(false);
   const reactFlowInstance = useReactFlow();
   const hasInitialized = useRef(false);
+  const [showNodePalette, setShowNodePalette] = useState(false);
 
   useEffect(() => {
     if (selectedNode !== null) setShowNodePalette(false);
@@ -380,14 +373,7 @@ const ReactFlowCanvasInner: React.FC<ReactFlowCanvasProps> = ({
     }
     return newPosition;
   };
-  const createNode = (
-    type: string,
-    category: "triggers" | "actions" | "conditions",
-  ) => {
-    const nodeType = category.slice(0, -1) as
-      | "trigger"
-      | "action"
-      | "condition";
+  const createNode = (type: string, nodeType: NodeType) => {
     let defaultConfig = {};
 
     const newPosition = getNewNodePosition();
@@ -397,8 +383,8 @@ const ReactFlowCanvasInner: React.FC<ReactFlowCanvasProps> = ({
       type: nodeType,
       position: newPosition,
       data: { [nodeType + "Type"]: type, config: defaultConfig },
-      inputs: category === "triggers" ? [] : ["input"],
-      outputs: category === "conditions" ? ["true", "false"] : ["output"],
+      inputs: nodeType === "trigger" ? [] : ["input"],
+      outputs: nodeType === "condition" ? ["true", "false"] : ["output"],
     };
 
     const updatedNodes = [...workflowNodes, newNode];
@@ -441,35 +427,6 @@ const ReactFlowCanvasInner: React.FC<ReactFlowCanvasProps> = ({
       duration: 300,
       interpolate: "linear",
     });
-  };
-
-  // Helper function to get node categories from registries
-  const getNodeCategories = () => {
-    // Initialize registries to ensure they're loaded
-    initializeTriggers();
-    initializeActions();
-
-    const actionRegistry = ActionRegistry.getInstance();
-    const triggerRegistry = TriggerRegistry.getInstance();
-
-    const categories = {
-      triggers: triggerRegistry.getAllTriggerMetadata().map((metadata) => ({
-        type: metadata.type,
-        label: metadata.label,
-        icon: metadata.icon,
-        description: metadata.description,
-      })),
-      actions: actionRegistry.getAllActionMetadata().map((metadata) => ({
-        type: metadata.type,
-        label: metadata.label,
-        icon: metadata.icon,
-        description: metadata.description,
-      })),
-      // TODO: Add conditions when we have a condition registry
-      conditions: [],
-    };
-
-    return categories;
   };
 
   // Define custom node types
@@ -575,81 +532,7 @@ const ReactFlowCanvasInner: React.FC<ReactFlowCanvasProps> = ({
       </Tooltip>
 
       {/* Add Node Button */}
-      <ActionIcon
-        style={{ position: "absolute", top: 16, right: 16 }}
-        onClick={() => setShowNodePalette(true)}
-      >
-        <IconPlus size={32} />
-      </ActionIcon>
-      {/* Drawer */}
-      <Transition
-        mounted={showNodePalette}
-        transition="slide-left"
-        duration={200}
-        timingFunction="ease"
-      >
-        {(styles) => (
-          <Paper
-            shadow="md"
-            p="md"
-            style={{
-              ...styles,
-              position: "absolute",
-              top: 0,
-              right: 0,
-              height: "100%",
-              width: 300,
-              zIndex: 1,
-            }}
-          >
-            <Text fw={500} mb="md">
-              Add Node
-            </Text>
-            <ScrollArea h="100%" pb="xl">
-              <Stack>
-                {Object.entries(getNodeCategories()).map(
-                  ([category, items]) => {
-                    if (items.length === 0) return null;
-                    return (
-                      <div key={category}>
-                        <Text
-                          size="sm"
-                          fw={500}
-                          c="dimmed"
-                          tt="capitalize"
-                          mb="xs"
-                        >
-                          {category}
-                        </Text>
-                        {items.map((item) => (
-                          <Button
-                            key={item.type}
-                            variant="subtle"
-                            fullWidth
-                            justify="start"
-                            leftSection={<Text size="lg">{item.icon}</Text>}
-                            mb="xs"
-                            onClick={() =>
-                              createNode(item.type, category as any)
-                            }
-                          >
-                            <Stack align="flex-start" gap={0}>
-                              <Text size="sm">{item.label}</Text>
-                              <Text size="xs" c="dimmed">
-                                {item.description}
-                              </Text>
-                            </Stack>
-                          </Button>
-                        ))}
-                      </div>
-                    );
-                  },
-                )}
-              </Stack>
-            </ScrollArea>
-          </Paper>
-        )}
-      </Transition>
+      <AddNodeList createNode={createNode} opened={showNodePalette} />
     </Box>
   );
 };
